// packages/theme/src/types.ts
export interface ColorPalette {
  50: string
  100: string
  200: string
  300: string
  400: string
  500: string
  600: string
  700: string
  800: string
  900: string
  950: string
}

export interface ThemeConfig {
  slug: string
  name: string
  primaryColor: ColorPalette
  accentColor: ColorPalette
  logo: {
    light: string
    dark?: string
    icon?: string
  }
  fonts: {
    primary: string
    heading: string
  }
  meta: {
    title: string
    description: string
    keywords: string[]
  }
}